fix visibility and coherency with fractal

+++ libs/messages/src/point.rs
@@ -1,21 +1,21 @@
 pub struct Point {
-    x: f32,
-    y: f32,
+    x: f64,
+    y: f64,
 }
 
 impl Point {
-    pub fn new(x: f32, y: f32) -> Option<Point> {
+    pub fn new(x: f64, y: f64) -> Option<Point> {
         if x.is_nan() || y.is_nan() {
             return None;
         }
         Some(Point { x, y })
     }
 
-    pub fn x(&self) -> f32 {
+    pub fn x(&self) -> f64 {
         self.x
     }
 
-    pub fn y(&self) -> f32 {
+    pub fn y(&self) -> f64 {
         self.y
     }
 }
